fix: integration with client
