+++ packages/react-components/src/lib/index.ts
@@ -10,3 +10,4 @@ export * from './cssbaseline';
 export * from './theme';
 export * from './cesium-map';
 export * from './autocomplete';
+export * from './file-picker';
